Rename package, update test framework
