+++ test_requirements.txt
@@ -1,2 +1,2 @@
-pytest==3.6.2
+pytest==5.2.0
 mock==2.0.0; python_version < '3.4'
